model is medium complete.

+++ server/model/vote_log.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type VoteLog struct {
-	UserID       int       `json:"user_id" db:"user_id"`
-	ImageID      int       `json:"image_id" db:"image_id"`
-	TagID        int       `json:"tag_id" db:"tag_id"`
+	UserID       int64     `json:"user_id" db:"user_id"`
+	ImageID      int64     `json:"image_id" db:"image_id"`
+	TagID        int64     `json:"tag_id" db:"tag_id"`
 	TimestampUTC time.Time `json:"timestamp_utc" db:"timestamp_utc"`
 	IsUpvote     bool      `json:"is_upvote" db:"is_upvote"`
 }
@@ -14,7 +14,7 @@ func (vl VoteLog) TableName() string {
 	return "vote_log"
 }
 
-func NewVoteLog(userID, imageID, tagID int, isUpvote bool) *VoteLog {
+func NewVoteLog(userID, imageID, tagID int64, isUpvote bool) *VoteLog {
 	return &VoteLog{
 		UserID:       userID,
 		ImageID:      imageID,
